Move handler tests
Closes #237.
