+++ internal/handlers/msg_delete_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package jsonb1
+package handlers
 
 import (
 	"testing"
@@ -21,21 +21,19 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/FerretDB/FerretDB/internal/types"
+	"github.com/FerretDB/FerretDB/internal/util/testutil"
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
 func TestDelete(t *testing.T) {
-	t.Parallel()
-
-	ctx, h, schema := setup(t)
+	ctx, h, pool := setup(t, nil)
+	schema := testutil.Schema(ctx, t, pool)
 
 	header := wire.MsgHeader{
 		OpCode: wire.OP_MSG,
 	}
 
 	t.Run(schema, func(t *testing.T) {
-		t.Parallel()
-
 		for i := 1; i <= 5; i++ {
 			var msg wire.OpMsg
 			err := msg.SetSections(wire.OpMsgSection{
@@ -138,8 +136,6 @@ func TestDelete(t *testing.T) {
 		for name, tc := range testCases {
 			tc := tc
 			t.Run(name, func(t *testing.T) {
-				t.Parallel()
-
 				tc.req.Set("$db", schema)
 
 				var reqMsg wire.OpMsg
